V.1.0.0 -> V.1.1.0 | fix: Action space is changed to only include available actions
Actions & values are stored in dictionary now and every time a decision is calculated it is added to this dictionary. This means that the size of dictionary changes as more actions are valid. This is more scalable and fixes errors where an action that cannot be completed was either calculated or initialised to some placeholder value.