V.1.1.0 -> V.1.2.0 | fix: Include 4th down stats in data processing
Fixes 'csv issue' as that issue was trying to access stats that were not available. There are no instances in the data for 4th down, field section 10, timeout used. I have included some code that skips over an action if there are no stats for it, but this will have to be handled with properly.

Moved functions for each action into their own file to declutter the model file and each action function.