V.1.4.1 -> V.1.5.0 | fix: Fourth down handling, first down usage fix, no data available fix
Some funky stuff was happening when handling fourth down. Wasn't properly considering when coming up short on 4th down the teams should switch off/def. Is now treated seperately to 1st, 2nd & 3rd downs.

Also changed first down info in State object to be the section that must be reached to get first down. In some places I had implemented this, in others I took in the distance from the current position to the first down section. Fixed the inconsistency.

Also sidestepped the issue of no data in for 4th down, ball in section 10, timeout called by deeming a timeout call as not a valid action. This should only be a temporary fix and should be investigated further, for that specific case and also all 4th down data (increase robustness of data)