V.1.6.2 -> V.1.7.0 | ordering of actions & outcomes changed to improve runtime
The order of actions changed and the value of previously calculated actions are remembered. Order of outcomes changed to go from worst outcome to best. After each outcome, the model checks if it is possible to reach the value of the optimal action. If not, the calculation of the current action is terminated.