V.2.2.0 -> V.2.3.0 | feat: early stopping conditions using current otpimal value
Created an upper bound on each action that gets updated as expected value of that action is calculatd. If that upper bound is less than the current optimal value, stop calculating that action

Cleaned up timeout handling so that main function doesn't need a specific input for delayed/immediate timeout

Still errs in code. Have TODO in some sections to denote this